requirements.txt
streamlit
pandas
plotly
google-genai
pydantic 
requests 
python-dotenv